+++ app/Services/BookServices.php
@@ -6,6 +6,11 @@ use GuzzleHttp\Client;
 use Vision\Feature;
 use Vision\Request\Image\LocalImage;
 use Vision\Vision;
+use Illuminate\Support\Facades\Mail;
+use Modules\User\Entities\User;
+use App\Models\KnowledgeCafe\Library\Book;
+use App\Models\Comment;
+use Illuminate\Support\Facades\DB;
 
 class BookServices
 {
@@ -22,7 +27,7 @@ class BookServices
         ]);
         $book = json_decode($res->getBody(), true);
 
-        if (! isset($book['items'])) {
+        if (!isset($book['items'])) {
             $res = $client->request('GET', 'https://www.googleapis.com/books/v1/volumes?q=ISBN:' . $isbn, [
                 'timeout' => 5.0,
             ]);
@@ -30,7 +35,7 @@ class BookServices
             $book = json_decode($res->getBody(), true);
         }
 
-        if (! isset($book['items'])) {
+        if (!isset($book['items'])) {
             return 'please try again';
         }
 
@@ -63,5 +68,35 @@ class BookServices
         $description = str_replace('-', '', trim($description));
 
         return $description;
+    }
+    public function SendMailUncommentedUsers()
+    {
+        $comment_id = comment::pluck('commentable_id');
+        $user_id = DB::table('book_readers')->whereNotIn('library_book_id', $comment_id)->pluck('user_id');
+        $book_id = DB::table('book_readers')->where('user_id', $user_id)->pluck('library_book_id');
+        $book_title = Book::wherenull('deleted_at')->orwhere('id', $book_id)->pluck('title')->toArray();
+        $user_name = USER::wherenull('deleted_at')->orwhere('id', $user_id)->pluck('name')->toArray();
+        $Id = [];
+        foreach (array_unique($user_name) as $key => $values) {
+            $temp = array_values(array_intersect_key($book_title, array_intersect($user_name, [$values])));
+            if (count($temp) > 1) {
+                $Id[$values] = $temp;
+            } else {
+                $Id[$values] = $temp[0];
+            }
+        }
+
+        foreach ($Id as $key => $value) {
+            $email = User::where('name', $key)->select('email')->pluck('email')->toArray();
+            foreach ($email as $mail) {
+                $data = ['name' => $key, 'key1' => $value];
+                Mail::send('emails.mailsend', $data, function ($message) use ($mail) {
+                    $message->to($mail);
+                    $message->subject('Feedback_on_Book');
+                });
+            }
+        }
+
+        return 0;
     }
 }
